+++ semcore/core/src/index.ts
@@ -1,0 +1,5 @@
+import createComponent from '@semcore/utils/lib/core';
+
+export default createComponent;
+
+export * from '@semcore/utils/lib/core';