+++ semcore/utils/src/addonTextChildren.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-// @ts-ignore
-import { CHILDREN_COMPONENT, INHERITED_NAME } from '@semcore/core';
+import { CHILDREN_COMPONENT, INHERITED_NAME } from './core';
 import getOriginChildren from './getOriginChildren';
 import isNode from './isNode';