+++ semcore/utils/src/core/enhancement/Root.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
-// @ts-ignore
-import assignProps from '@semcore/utils/lib/assignProps';
-// @ts-ignore
-import { useForkRef } from '@semcore/utils/lib/ref';
+import assignProps from '../../assignProps';
+import { useForkRef } from '../../ref';
 
 function createRootRender() {
   // @ts-ignore