+++ semcore/utils/src/core/styled/index.tsx
@@ -1,4 +1,4 @@
-import { UnknownProperties } from '..';
+import { UnknownProperties } from '../types/UnknownProperties';
 
 export type StyledProps = {
   // eslint-disable-next-line ssr-friendly/no-dom-globals-in-module-scope