+++ semcore/utils/src/core/types/index.ts
@@ -1,0 +1,5 @@
+import { UnknownProperties } from './UnknownProperties';
+
+export * from './Component';
+
+export { UnknownProperties };