+++ semcore/utils/src/core/types/symbols.ts
@@ -1,0 +1,7 @@
+export const CORE_COMPONENT = Symbol('CORE_COMPONENT');
+export const CORE_INSTANCE = Symbol('CORE_INSTANCE');
+export const CORE_AS_PROPS = Symbol('CORE_AS_PROPS');
+export const CORE_INIT = Symbol('CORE_INIT');
+export const CONTEXT_COMPONENT = Symbol('CONTEXT_COMPONENT');
+export const CREATE_COMPONENT = Symbol('CREATE_COMPONENT');
+export const PARENT_COMPONENTS = Symbol('PARENT_COMPONENTS');